# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 五目オンライン API
  version: 1.0.0
  description: |
    API for GomokuOnline.
    五目オンライン用API。
servers:
  - url: "https://gomokuonline.nodokamome.com/api/{ver}"
    variables:
      ver:
        enum:
          - v1
        default: v1
tags:
  - name: Player
    description: Manage players.
  - name: Version
    description: Manage versions.

security:
  - token: []

paths:
  "/players":
    get:
      tags:
        - Player
      summary: プレイヤー一覧
      parameters:
        - in: query
          name: player_id
          schema:
            type: string
            example: abcde123
        - in: query
          name: state
          schema:
            type: string
            enum:
              - online
              - offline
            example: online
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - rank
            example: rank
      responses:
        "200":
          description: |
            取得に成功しました
        "202":
          description: |
            プレイヤーが見つかりませんでした
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
    post:
      tags:
        - 
      summary: プレイヤーの新規登録
      requestBody:
        required: true
        content:
          appliction/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    playerId:
                      type: string
                      example: abcde123
                    name:
                      type: string
                      example: hoge
                    icon:
                      type: string
                      example: icon_1
      responses:
        "200":
          description: |
            登録に成功しました
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
  "/players/{player_id}/prof":
    put:
      tags:
        - Player
      summary: プレイヤーの情報更新
      parameters:
        - in: path
          name: player_id
          schema:
            type: string
            example: abcde123
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: hoge
                icon:
                  type: string
                  example: icon_1
      responses:
        "200":
          description: |
            登録に成功しました
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
  "/players/{player_id}/match_record":
    put:
      tags:
        - Player
      summary: 試合結果のインクリメント
      parameters:
        - in: path
          name: player_id
          schema:
            type: string
            example: abcde123
          required: true
      requestBody:
        required: true
        description: match_recordをwinにすれば勝利数が1インクリメント。loseにすれば敗北数1ディクリメントします。
        content:
          application/json:
            schema:
              type: object
              properties:
                match_record:
                  type: string
                  example: win
      responses:
        "200":
          description: |
            取得に成功しました
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
  "/players/{player_id}/state":
    put:
      tags:
        - Player
      summary: オンライン状態設定
      parameters:
        - in: path
          name: player_id
          schema:
            type: string
            example: abcde123
          required: true
      requestBody:
        required: true
        description: stateはonline or offlineを入力してください。
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  example: online
      responses:
        "200":
          description: |
            取得に成功しました
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
  "/versions":
    get:
      tags:
        - Version
      summary: バージョン一覧
      parameters:
        - in: query
          name: platform
          schema:
            type: string
            example: platform_name
            enum:
              - ios
              - android
          required: true
      responses:
        "200":
          description: |
            取得に成功しました
        "401":
          description: |
            認証に失敗しました
        "403":
          description: |
            エラー
components:
  securitySchemes:
    token:
      type: "apiKey"
      name: "Authorization"
      in: "header"